Added double preview windows

+++ src/model_loaders/CRLCameraModels.cpp
@@ -157,7 +157,7 @@ void CRLCameraModels::Model::setGrayscaleTexture(crl::multisense::image::Header 
         }
 
         for (int i = 0; i < streamOne->imageLength / 2; ++i) {
-            float intermediate = (float) ((float) p[i] / max) * 65536.0f;
+            float intermediate = (float) ((float) p[i] / (float) max) * 65536.0f;
             p[i] = (uint16_t) intermediate;
         }
